Remove empty section
It was highlighted to me that the [How to Add Markup](https://bioschemas.org/tutorials/howto/howto_add_markup) tutorial is incomplete and that the final section leaves an unfinished feel to the tutorial.

@aurel-l is down to check this tutorial in [Issue 488](https://github.com/BioSchemas/specifications/issues/488).

For now I've removed the section, but what is needed is a pointer to the [How to check your Bioschemas deployment](https://bioschemas.org/tutorials/howto/howto_check_deploy) tutorial. But before that there needs to be content on how to add your markup.